requirements.txt
docker-py==1.10.6
Flask==1.1.2
Flask-Cache==0.13.1
Flask-Caching==1.9.0
Flask-Limiter==1.4
uwsgi
